Refactor SnapshotReader and rename to SnapshotServerReader.
