+++ debezium-connector-mysql/src/main/java/io/debezium/connector/mysql/MySqlConnectorTask.java
@@ -146,10 +146,10 @@ public final class MySqlConnectorTask extends SourceTask {
             BinlogReader binlogReader = new BinlogReader("binlog", taskContext);
             if (startWithSnapshot) {
                 // We're supposed to start with a snapshot, so set that up ...
-                SnapshotReader snapshotReader = new SnapshotReader("snapshot", taskContext);
-                snapshotReader.useMinimalBlocking(taskContext.useMinimalSnapshotLocking());
-                if (snapshotEventsAreInserts) snapshotReader.generateInsertEvents();
-                readers.add(snapshotReader);
+                SnapshotServerReader snapshotServerReader = new SnapshotServerReader("snapshot", taskContext);
+                snapshotServerReader.useMinimalBlocking(taskContext.useMinimalSnapshotLocking());
+                if (snapshotEventsAreInserts) snapshotServerReader.generateInsertEvents();
+                readers.add(snapshotServerReader);
 
                 if (taskContext.isInitialSnapshotOnly()) {
                     logger.warn("This connector will only perform a snapshot, and will stop after that completes.");
