+++ debezium-connector-mysql/src/test/resources/log4j.properties
@@ -11,5 +11,5 @@ log4j.rootLogger=WARN, stdout
 log4j.logger.io.debezium=WARN
 log4j.logger.io.debezium.embedded.EmbeddedEngine$EmbeddedConfig=WARN
 #log4j.logger.io.debezium.connector.mysql.BinlogReader=DEBUG
-#log4j.logger.io.debezium.connector.mysql.SnapshotReader=DEBUG
+#log4j.logger.io.debezium.connector.mysql.SnapshotServerReaderReader=DEBUG
 #log4j.logger.io.debezium.relational.history=DEBUG
